graceful exit and garbage collection at beginning
