+++ go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/imdario/mergo v0.3.11 // indirect
 	github.com/libvirt/libvirt-go v6.9.0+incompatible
 	github.com/mitchellh/copystructure v1.0.0 // indirect
+	github.com/mitchellh/go-ps v1.0.0
 	github.com/olekukonko/tablewriter v0.0.4
 	github.com/vishvananda/netlink v1.1.0
 	golang.org/x/crypto v0.0.0-20201016220609-9e8e0b390897
